# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ChronosGuard Compliance API",
    "description": "Temporal Compliance RAG Engine \u2014 audits internal policies against in-force regulation for a jurisdiction as of a given date.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/audits": {
      "post": {
        "tags": [
          "audits"
        ],
        "summary": "Create Audit",
        "operationId": "create_audit",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuditCreate"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditRunOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "audits"
        ],
        "summary": "List Audits",
        "operationId": "list_audits",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_AuditRunOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/audits/{run_id}": {
      "get": {
        "tags": [
          "audits"
        ],
        "summary": "Get Audit",
        "operationId": "get_audit",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Run Id"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditRunOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/audits/{run_id}/findings": {
      "get": {
        "tags": [
          "audits"
        ],
        "summary": "List Audit Findings",
        "operationId": "list_audit_findings",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Run Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_FindingOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AuditCreate": {
        "properties": {
          "policy_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Policy Id"
          },
          "policy_text": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200000,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Policy Text"
          },
          "jurisdiction": {
            "type": "string",
            "maxLength": 16,
            "minLength": 2,
            "title": "Jurisdiction"
          },
          "as_of_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of Date"
          }
        },
        "type": "object",
        "required": [
          "jurisdiction"
        ],
        "title": "AuditCreate",
        "example": {
          "as_of_date": "2026-06-06",
          "jurisdiction": "PK",
          "policy_id": 1
        }
      },
      "AuditRunOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "succeeded",
              "partial",
              "failed"
            ],
            "title": "Status"
          },
          "verdict": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "COMPLIANT",
                  "VIOLATIONS_FOUND",
                  "INSUFFICIENT_EVIDENCE"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Verdict"
          },
          "coverage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Coverage"
              },
              {
                "type": "null"
              }
            ]
          },
          "stale": {
            "type": "boolean",
            "title": "Stale"
          },
          "jurisdiction": {
            "type": "string",
            "title": "Jurisdiction"
          },
          "as_of_date": {
            "type": "string",
            "format": "date",
            "title": "As Of Date"
          },
          "policy_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Policy Id"
          },
          "policy_version_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Policy Version Id"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "total_tokens": {
            "type": "integer",
            "title": "Total Tokens"
          },
          "cost_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Usd"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "verdict",
          "coverage",
          "stale",
          "jurisdiction",
          "as_of_date",
          "policy_id",
          "policy_version_id",
          "model",
          "total_tokens",
          "cost_usd",
          "error",
          "created_at",
          "finished_at"
        ],
        "title": "AuditRunOut"
      },
...
      "Coverage": {
        "properties": {
          "violation": {
            "type": "integer",
            "title": "Violation",
            "default": 0
          },
          "compliant": {
            "type": "integer",
            "title": "Compliant",
            "default": 0
          },
          "insufficient_evidence": {
            "type": "integer",
            "title": "Insufficient Evidence",
            "default": 0
          },
          "error": {
            "type": "integer",
            "title": "Error",
            "default": 0
          }
        },
        "type": "object",
        "title": "Coverage"
      },
...
      "FindingOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "clause_index": {
            "type": "integer",
            "title": "Clause Index"
          },
          "risk_level": {
            "type": "string",
            "enum": [
              "HIGH",
              "MEDIUM",
              "LOW"
            ],
            "title": "Risk Level"
          },
          "offending_policy_text": {
            "type": "string",
            "title": "Offending Policy Text"
          },
          "legal_rule_text": {
            "type": "string",
            "title": "Legal Rule Text"
          },
          "citation": {
            "type": "string",
            "title": "Citation"
          },
          "grounding_quote": {
            "type": "string",
            "title": "Grounding Quote"
          },
          "rationale": {
            "type": "string",
            "title": "Rationale"
          },
          "suggested_fix": {
            "type": "string",
            "title": "Suggested Fix"
          },
          "source_chunk_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Chunk Id"
          },
          "source_document_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Document Id"
          },
          "source_url": {
            "type": "string",
            "title": "Source Url"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "needs_review": {
            "type": "boolean",
            "title": "Needs Review"
          }
        },
        "type": "object",
        "required": [
          "id",
          "clause_index",
          "risk_level",
          "offending_policy_text",
          "legal_rule_text",
          "citation",
          "grounding_quote",
          "rationale",
          "suggested_fix",
          "source_chunk_id",
          "source_document_id",
          "source_url",
          "confidence",
          "needs_review"
        ],
        "title": "FindingOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Page_AuditRunOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AuditRunOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[AuditRunOut]"
      },
...
      "Page_FindingOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/FindingOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[FindingOut]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}